requirements.txt
streamlit
google-generativeai
psycopg2-binary
gitpython
sentence_transformers
cocoindex
